+++ src/lang/program.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
 #include <array>
+#include <map>
+#include <string>
 #include <vector>
 
 #include "lang/number.hpp"
@@ -129,4 +131,5 @@ class Program {
   bool operator<(const Program &p) const;
 
   std::vector<Operation> ops;
+  std::map<std::string, int64_t> directives;
 };